dbhub: Fix size for very large databases in remote dock
Fix the size value in the database list of the Remote Dock for very
large databases. This was showing a negative value due to an overlow.

+++ src/RemoteModel.cpp
@@ -89,7 +89,7 @@ std::vector<RemoteModelItem*> RemoteModelItem::loadArray(const json& array, Remo
         if(elem.contains("commit_id"))
             item->setValue(RemoteModelColumnCommitId, QString::fromStdString(elem["commit_id"]));
         if(elem.contains("size"))
-            item->setValue(RemoteModelColumnSize, QString::number(static_cast<int>(elem["size"])));
+            item->setValue(RemoteModelColumnSize, QString::number(static_cast<unsigned long>(elem["size"])));
 
         items.push_back(item);
     }
